Web_Crawler/for_imgur/requirements.txt
Pillow
requests
bs4